pyproject.toml
[project]
name = "prometheus_humidifier_exporter"
version = "1.0.0"
authors = [
    { name="Georgi Todorov", email="georgi.vladimirov.todorov@gmail.com" },
]
description = "Prometheus exporter for Xiaomi Smartmi Air Humidifier 2"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "python-miio==0.4.7",
    "prometheus_client==0.19.0",
    "build==1.0.3"
]

[project.scripts]
app = "run_server:main"

[project.urls]
Homepage = "https://github.com/byumov/prometheus_humidifier_exporter"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"